# backend/service/agent-invoice/openapi.yaml
openapi: "3.1.0"
info:
  title: agent-invoice-service
  version: "1.1.0"
  description: Invoice & Agent Fee Management Service — agents, agent fees, and agent invoice operations
servers:
  - url: http://agent-invoice-service:3102
    description: Internal service URL
security:
  - gatewaySecret: []
...
components:
  securitySchemes:
    gatewaySecret:
      type: apiKey
      in: header
      name: x-gateway-secret
  parameters:
    xUserOu:
      name: x-user-ou
      in: header
      required: true
      schema:
        type: string
    xUserBranch:
      name: x-user-branch
      in: header
      required: true
      schema:
        type: string
    xUserId:
      name: x-user-id
      in: header
      required: true
      schema:
        type: string
    xUserRole:
      name: x-user-role
      in: header
      required: true
      schema:
        type: string
    xRequestId:
      name: x-request-id
      in: header
      required: false
      schema:
        type: string
...
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, code, message, data, requestId]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
        data:
          type: "null"
        requestId:
          type: string
    Agent:
      type: object
      properties:
        _id:
          type: string
        ou_id:
          type: string
        branch_id:
          type: string
        branch_code:
          type: string
        branch_name:
          type: string
        branch_type:
          type: string
          enum: [MA, AG]
        branch_desc:
          type: string
        parent_branch_id:
          type: string
        ref_fee_branch_id:
          type: string
          nullable: true
        ref_fee_branch_name:
          type: string
          nullable: true
        currency:
          type: string
        default_fee_rate:
          type: number
          nullable: true
        active:
          type: boolean
        cr_by:
          type: string
        cr_date:
          type: string
          format: date-time
        cr_prog:
          type: string
        upd_by:
          type: string
        upd_date:
          type: string
          format: date-time
        upd_prog:
          type: string
...
    ApiEnvelope:
      type: object
      required: [success, code, message, data]
      properties:
        success:
          type: boolean
        code:
          type: string
        message:
          type: string
        data:
          nullable: true
        requestId:
          type: string
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              const: false
            data:
              nullable: true
    AgentInvoiceListItem:
      type: object
      description: List projection includes cr_date only; other audit fields are omitted from list responses.
      required: [_id, ou_id, branch_id, iv_no, status, cr_date]
      properties:
        _id:
          type: string
        ou_id:
          type: string
        branch_id:
          type: string
        iv_no:
          type: string
        net_win:
          type: number
          nullable: true
        amount:
          type: number
          nullable: true
        branch_name:
          type: string
          nullable: true
        billing_month:
          type: string
          nullable: true
        due_date:
          type: string
          description: Payment due date (YYYY-MM-15 of the invoice cycle month)
          example: "2026-06-15"
          nullable: true
        status:
          type: string
          enum: [PENDING, VOID, CAL, MISSING_FEE, READY, ERROR, PAID]
        cr_date:
          type: string
          format: date-time
    AgentInvoiceDetail:
      allOf:
        - $ref: "#/components/schemas/AgentInvoiceListItem"
        - type: object
          required: [ou_name, cr_by, cr_prog, upd_by, upd_prog, upd_date]
          properties:
            ou_name:
              type: string
              nullable: true
            cr_by:
              type: string
            cr_prog:
              type: string
            upd_by:
              type: string
            upd_prog:
              type: string
            upd_date:
              type: string
              format: date-time
    AgentInvoiceTransaction:
      type: object
      required:
        - _id
        - ref_iv_id
        - ou_id
        - branch_id
        - company_id
        - main_category_id
        - net_win
        - fee
        - amount
        - cr_by
        - cr_prog
        - cr_date
        - upd_by
        - upd_prog
        - upd_date
      properties:
        _id:
          type: string
        ref_iv_id:
          type: string
        ou_id:
          type: string
        ou_name:
          type: string
          nullable: true
        branch_id:
          type: string
        branch_name:
          type: string
          nullable: true
        company_id:
          type: string
        company_name:
          type: string
          nullable: true
        main_category_id:
          type: string
        main_category_name:
          type: string
          nullable: true
        net_win:
          type: number
        fee:
          oneOf:
            - type: number
            - type: string
              const: N/A
        amount:
          type: number
        cr_by:
          type: string
        cr_prog:
          type: string
        cr_date:
          type: string
          format: date-time
        upd_by:
          type: string
        upd_prog:
          type: string
        upd_date:
          type: string
          format: date-time
...
    UpdateStatusRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [PAID]
...
    InvoiceAgentBranch:
      type: object
      required: [branch_id, branch_name, branch_code]
      properties:
        branch_id:
          type: string
          minLength: 24
          maxLength: 24
          pattern: "^[a-f0-9]{24}$"
        branch_name:
          type: string
          nullable: true
        branch_code:
          type: string
          nullable: true
    ListInvoiceAgentsSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              const: true
            data:
              type: array
              items:
                $ref: "#/components/schemas/InvoiceAgentBranch"
    ListInvoicesSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              const: true
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: "#/components/schemas/AgentInvoiceListItem"
                pagination:
                  type: object
                  required: [page, limit, total, totalPages]
                  properties:
                    total:
                      type: integer
                      nullable: true
                      description: Total matching documents; null when count is unavailable (see hasMore).
                    page:
                      type: integer
                    limit:
                      type: integer
                    totalPages:
                      type: integer
                      nullable: true
                      description: Derived from total; null when total is null.
                    hasMore:
                      type: boolean
                      description: Present only when total is null; indicates another page may exist.
    ListTransactionsSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              const: true
            data:
              type: array
              items:
                $ref: "#/components/schemas/AgentInvoiceTransaction"
...
    GetDetailSuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          properties:
            success:
              const: true
            data:
              $ref: "#/components/schemas/AgentInvoiceDetail"
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Gateway authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Missing or invalid user context
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: Internal error (includes DATASTORE_CREDENTIAL_REJECTED on MongoDB auth failure)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /api/v1/agent-invoice/agents:
    get:
      tags: [agents]
      summary: List agents with pagination
      parameters:
        - $ref: "#/components/parameters/xUserOu"
        - $ref: "#/components/parameters/xUserBranch"
        - $ref: "#/components/parameters/xUserId"
        - $ref: "#/components/parameters/xUserRole"
        - $ref: "#/components/parameters/xRequestId"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  code:
                    type: string
                    example: SUCCESS
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Agent"
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    post:
      tags: [agents]
      summary: Create a new agent
      parameters:
        - $ref: "#/components/parameters/xUserOu"
        - $ref: "#/components/parameters/xUserBranch"
        - $ref: "#/components/parameters/xUserId"
        - $ref: "#/components/parameters/xUserRole"
        - $ref: "#/components/parameters/xRequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                [branch_id, branch_code, branch_name, branch_type, currency]
              properties:
                branch_id:
                  type: string
                  pattern: "^[0-9a-fA-F]{24}$"
                branch_code:
                  type: string
                branch_name:
                  type: string
                branch_type:
                  type: string
                  enum: [MA, AG]
                branch_desc:
                  type: string
                parent_branch_id:
                  type: string
                  pattern: "^[0-9a-fA-F]{24}$"
                currency:
                  type: string
                default_fee_rate:
                  type: number
                  minimum: 0
                  maximum: 100
      responses:
        "201":
          description: Created
          headers:
            ETag:
              schema:
                type: string
              description: "Weak ETag for optimistic locking"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  code:
                    type: string
                    example: CREATED
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      insertedId:
                        type: string
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /api/v1/invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices with filters
      description: |
        Paginated invoice list filtered by optional query params (iv_no, branch_id, billing_month, status).
        Results are scoped to `ou_id` matching `x-user-ou`.
      parameters:
        - $ref: "#/components/parameters/xUserOu"
        - $ref: "#/components/parameters/xUserBranch"
        - $ref: "#/components/parameters/xUserId"
        - $ref: "#/components/parameters/xUserRole"
        - $ref: "#/components/parameters/xRequestId"
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: iv_no
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: string
            minLength: 24
            maxLength: 24
            pattern: "^[a-f0-9]{24}$"
        - name: billing_month
          in: query
          schema:
            type: string
            pattern: '^\\d{4}-(0[1-9]|1[0-2])$'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, VOID, CAL, MISSING_FEE, READY, ERROR, PAID]
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListInvoicesSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/invoices/agent:
    get:
      operationId: listInvoiceAgents
      tags: [invoices]
      summary: List branches for invoice operations
      description: |
        Returns branches from `gpp_777ww.su_branch` filtered by `ou_id` matching `x-user-ou`.
      parameters:
        - $ref: "#/components/parameters/xUserOu"
        - $ref: "#/components/parameters/xUserBranch"
        - $ref: "#/components/parameters/xUserId"
        - $ref: "#/components/parameters/xUserRole"
        - $ref: "#/components/parameters/xRequestId"
      responses:
        "200":
          description: Branch list for the user's OU
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListInvoiceAgentsSuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/invoices/{id}/transactions:
    get:
      operationId: listInvoiceTransactions
      tags: [invoices]
      summary: List invoice transactions
      description: |
        Returns agent_invoice_transaction rows for the invoice (ref_iv_id = path id) with display-name joins.
        Invoice must belong to `ou_id` matching `x-user-ou`.
      parameters:
        - $ref: "#/components/parameters/xUserOu"
        - $ref: "#/components/parameters/xUserBranch"
        - $ref: "#/components/parameters/xUserId"
        - $ref: "#/components/parameters/xUserRole"
        - $ref: "#/components/parameters/xRequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            pattern: "^[a-f0-9]{24}$"
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListTransactionsSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/invoices/{id}/status:
    put:
      operationId: putInvoiceStatus
      tags: [invoices]
      summary: Mark invoice as PAID (READY → PAID only)
      description: |
        Body must be status PAID only. Transition READY to PAID is allowed.
        Invoice must belong to `ou_id` matching `x-user-ou`.
      parameters:
        - $ref: "#/components/parameters/xUserOu"
        - $ref: "#/components/parameters/xUserBranch"
        - $ref: "#/components/parameters/xUserId"
        - $ref: "#/components/parameters/xUserRole"
        - $ref: "#/components/parameters/xRequestId"
        - name: id
          in: path
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            pattern: "^[a-f0-9]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetDetailSuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...